+++ internal/pkg/cache/get.go
@@ -60,7 +60,7 @@ type GetOpts struct {
 
 // RegistryPath fulfills the cacheOperationProvider interface for GetOpts
 func (opts *GetOpts) RegistryPath() string {
-	return path.Join(opts.cachePath, opts.RegistryName)
+	return path.Join(opts.cachePath, opts.RegistryName, opts.Ref)
 }
 
 // PackPath fulfills the cacheOperationProvider interface for GetOpts